ai_call_backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
websockets==12.0
anthropic==0.25.0
supabase==2.4.2
firebase-admin==6.4.0
python-dotenv==1.0.1
pydantic==2.6.4
httpx==0.27.0
python-multipart==0.0.9
deepgram-sdk==3.2.7
elevenlabs==1.2.0
pydub==0.25.1
audioop-lts==0.2.1    # CRITICAL: replaces removed audioop in Python 3.13
numpy==1.26.4
